가게 목록보기 기능 구현 및 테이블 재설계 (#25)
* 회원 비활성화 기능

* 계정 비활성화 컨트롤러, 서비스 코드구현 완료

* 회원 비활성화 Junit 테스트 코드 작성완료

* customerReview 엔티티 수정 및 DummyEntity 설정

* 리뷰 등록하기 기능 Controller, Service 구현

* 테스트 코드 작성 및 정상작동 확인

* h2 접근시 엔티티 기본생성자 접근오류 수정

* Menu DummyEntity 설계

* 더미데이터 및 oreder엔티티 정리

* 테이블 관계간 onetoone 어노테이션 부여

* 테이블연간관계 어노테이션 다시 설정

* 주문하기 기능하다가 집감

* order, order_detail 엔티티 설계수정

* 주말 작업중 점심식사하러감

* order분야 잠시 정리

* 내 리뷰 목록보기 기능 컨트롤러 서비스 코드 작성

* customerReview 엔티티수정 및 dto수정 설계

* ceoReview 더미데이터 설계

* 더미데이터 및 dto설계 중간저장

* 레파지토리 쿼리 작성 중 테이블 설계 피드백받기로함

* Review관련 엔티티 어노테이션 정리

* 테스트 코드에서 CeoReview가 바인딩 안되서 리스트 값이 0으로 뜸

* DummyEntity에 강제로 ceoReview값 부여 후 테스트 코드 성공

* 내 리뷰 목록보기 기능구현 완료 (#18)

* 내 리뷰 목록보기 기능 컨트롤러 서비스 코드 작성

* customerReview 엔티티수정 및 dto수정 설계

* ceoReview 더미데이터 설계

* 더미데이터 및 dto설계 중간저장

* 레파지토리 쿼리 작성 중 테이블 설계 피드백받기로함

* Review관련 엔티티 어노테이션 정리

* 테스트 코드에서 CeoReview가 바인딩 안되서 리스트 값이 0으로 뜸

* DummyEntity에 강제로 ceoReview값 부여 후 테스트 코드 성공

* 내 리뷰 비활성화하기 서비스, 컨트롤러, 테스트코드 작성 및 Devinit 파일에 더미데이터 삽입

* 내 리뷰 비활성화하기 기능 테스트 코드까지 통과완료

* 리뷰 숨기기 기능 구현 (#19)

* 내 리뷰 목록보기 기능 컨트롤러 서비스 코드 작성

* customerReview 엔티티수정 및 dto수정 설계

* ceoReview 더미데이터 설계

* 더미데이터 및 dto설계 중간저장

* 레파지토리 쿼리 작성 중 테이블 설계 피드백받기로함

* Review관련 엔티티 어노테이션 정리

* 테스트 코드에서 CeoReview가 바인딩 안되서 리스트 값이 0으로 뜸

* DummyEntity에 강제로 ceoReview값 부여 후 테스트 코드 성공

* 내 리뷰 비활성화하기 서비스, 컨트롤러, 테스트코드 작성 및 Devinit 파일에 더미데이터 삽입

* 내 리뷰 비활성화하기 기능 테스트 코드까지 통과완료

* DummyEntity, DevInit 코드 수정

* 해당 가게에 맞는 리뷰 불러오기 서비스코드까지 구현 but 별점평균 쿼리로 바꿔야할듯

* 해당가게에 맞는 리뷰 별점평균 쿼리문 작성완료

* 가게목록보기 기능 서비스코드 및 dto 설계

* 가게 목록보기 서비스코드, 컨트롤러, DTO  1차 설계 완료 DTO 응답 수정해달래서 수정하기전 커밋

* 임시코드 작성 집가서 복습

* 고객리뷰 등록하기 엔티티수정으로 인한 코드수정

* starPoint int->Double로 변경

* 가게 목록보기 기능 컨트롤러, 서비스, dto 구현완료

* 테스트 코드 작동 중 오류

* 가게 목록보기 기능 테스트 코드 통과 및 쿼리문 수정

* 테이블 재설계로 인한 더미데이터 변경 및 junit 테스트 코드 재설계요망

* 테이블 재설계 및 가게 목록보기 기능 구현 (#24)

* 내 리뷰 목록보기 기능 컨트롤러 서비스 코드 작성

* customerReview 엔티티수정 및 dto수정 설계

* ceoReview 더미데이터 설계

* 더미데이터 및 dto설계 중간저장

* 레파지토리 쿼리 작성 중 테이블 설계 피드백받기로함

* Review관련 엔티티 어노테이션 정리

* 테스트 코드에서 CeoReview가 바인딩 안되서 리스트 값이 0으로 뜸

* DummyEntity에 강제로 ceoReview값 부여 후 테스트 코드 성공

* 내 리뷰 비활성화하기 서비스, 컨트롤러, 테스트코드 작성 및 Devinit 파일에 더미데이터 삽입

* 내 리뷰 비활성화하기 기능 테스트 코드까지 통과완료

* DummyEntity, DevInit 코드 수정

* 해당 가게에 맞는 리뷰 불러오기 서비스코드까지 구현 but 별점평균 쿼리로 바꿔야할듯

* 해당가게에 맞는 리뷰 별점평균 쿼리문 작성완료

* 가게목록보기 기능 서비스코드 및 dto 설계

* 가게 목록보기 서비스코드, 컨트롤러, DTO  1차 설계 완료 DTO 응답 수정해달래서 수정하기전 커밋

* 임시코드 작성 집가서 복습

* 고객리뷰 등록하기 엔티티수정으로 인한 코드수정

* starPoint int->Double로 변경

* 가게 목록보기 기능 컨트롤러, 서비스, dto 구현완료

* 테스트 코드 작동 중 오류

* 가게 목록보기 기능 테스트 코드 통과 및 쿼리문 수정

* 테이블 재설계로 인한 더미데이터 변경 및 junit 테스트 코드 재설계요망

+++ src/main/java/shop/mtcoding/finalproject/dto/customerReview/CustomerReviewRespDto.java
@@ -15,7 +15,7 @@ public class CustomerReviewRespDto {
     @Setter
     public static class InsertCustomerReviewRespDto {
         private String content;
-        private int starPoint;
+        private Double starPoint;
         private String photo;
 
         public InsertCustomerReviewRespDto(CustomerReview customerReview) {
@@ -73,7 +73,7 @@ public class CustomerReviewRespDto {
         public class CustomerReviewDto {
             private String content;
             private String photo;
-            private int starPoint;
+            private Double starPoint;
             private String createdAt;
             private String comment;
             private String commentCreatedAt;
@@ -96,15 +96,15 @@ public class CustomerReviewRespDto {
 // public static class CustomerReviewListRespDto {
 // private UserDto user;
 // private List<CustomerReviewDto> customerReviews = new ArrayList<>();
-// private List<CeoReviewDto> ceoReviews = new ArrayList<>();
+// private List<CeoReviewDto> CeoReview = new ArrayList<>();
 
 // public CustomerReviewListRespDto(User user, List<CustomerReview>
-// customerReviews, List<CeoReview> ceoReviews) {
+// customerReviews, List<CeoReview> CeoReview) {
 // this.user = new UserDto(user);
 // this.customerReviews = customerReviews.stream()
 // .map((customerReview) -> new CustomerReviewDto(customerReview))
 // .collect(Collectors.toList());
-// this.ceoReviews = ceoReviews.stream()
+// this.CeoReview = CeoReview.stream()
 // .map((ceoReview) -> new CeoReviewDto(ceoReview))
 // .collect(Collectors.toList());
 // }
@@ -127,7 +127,7 @@ public class CustomerReviewRespDto {
 // private Long id;
 // private String content;
 // private String photo;
-// private int starPoint;
+// private Double starPoint;
 // private String createAt;
 
 // public CustomerReviewDto(CustomerReview customerReview) {
